fix(multisig_quorum_animation): KeySafeAnimation widget 분리
